Many changes to regexs, colors, comments.
Never intented to upload, so there are many changes to the syntax file,
including colors, regexes, comments in this initial commit to fork.  This
fork is following a different aesthetic and goal, so probably no one is
interested in merging back, also orignal hasn't been updated, etc.  Will
try to explain in readme eventually.